# docs/swagger.yaml
basePath: /api/v1
definitions:
  api-rbac_internal_model.AssignPermissionsRequest:
    properties:
      permission_ids:
        description: '例: [1, 2, 3]'
        items:
          type: integer
        type: array
    required:
    - permission_ids
    type: object
...
  api-rbac_internal_model.CreateRoleRequest:
    properties:
      description:
        example: 负责内容编辑的角色
        maxLength: 255
        type: string
      name:
        example: 编辑员
        maxLength: 64
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  api-rbac_internal_model.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      updated_at:
        type: string
    type: object
  api-rbac_internal_model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/api-rbac_internal_model.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  api-rbac_internal_model.UpdateRoleRequest:
    properties:
      description:
        example: 负责高级内容编辑的角色
        maxLength: 255
        type: string
      name:
        example: 高级编辑员
        maxLength: 64
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  api-rbac_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  api-rbac_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: 通用的 RBAC 权限管理微服务，支持用户/角色/权限的 CRUD 及绑定关系管理。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC 权限管理系统 API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: 分页查询角色列表，支持关键词搜索
      parameters:
      - default: 1
        description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        example: 10
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词(角色名/描述)
        example: 管理员
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_pkg_response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 新增一个角色
      parameters:
      - description: 创建角色参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_internal_model.Role'
              type: object
        "400":
          description: 参数错误或角色名已存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /roles/{id}:
    delete:
      description: 根据 ID 删除一个角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    get:
      description: 根据 ID 获取角色信息，包含关联的权限列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_internal_model.Role'
              type: object
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新角色的名称和描述
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - 角色管理
  /roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: 覆盖式更新角色的权限列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 为角色分配权限
      tags:
      - 角色管理
  /roles/{id}/permissions/{permId}:
    delete:
      description: 移除角色的某个权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID
        in: path
        name: permId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 移除角色权限
      tags:
      - 角色管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"